Set up codebase to be installed with / managed by `uv`
This commit updates README.md instructions for use of `uv`. It
describes briefly how to manage dependencies in CONTRIBUTING.md.
It also adds `pyproject.toml` with deps transferred over from
`requirements.txt`.

Additionally, it deletes `requirements.txt`, and uses `uv` in
the sagemaker Dockerfile.
Currently in the Dockerfile it installs from `pyproject.toml`
instead of `uv.lock` because `uv.lock` seems to rely on the
sagemaker install tar file even when you're not installing it.